# qwixx_openapi.yml
openapi: 3.1.0
info:
  title: Qwixx Game API
  version: 1.0.0
  description: |
    OpenApi definition for the Qwixx dice game.
    This API manages players, games, moves, and game state.
servers:
  - url: http://localhost:4300
    description: Local development server
paths:
  /games:
    get:
      summary: List all active games
      description: Get stats like number of players, max nr players, whether the game is in progress.
      operationId: getAllGames
      tags:
        - games
      responses:
        '200':
          description: A list of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameInfo'
    post:
      summary: Create a new game
      description: |
        To start playing:
        1. Create a game
        2. Add players
        3. Start the game
      operationId: createNewGame
      tags:
        - games
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGameRequest'
      responses:
        '201':
          description: Game created
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
        '409':
          description: Game already exists with that name
  /games/{sessionId}:
    get:
      summary: Get details of a specific game
      operationId: getGame
      tags:
        - games
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Game details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameInfo'
        '404':
          description: Game not found
    post:
      summary: Start a game
      operationId: startGame
      tags:
        - games
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                botProfilePics:
                  type: array
                  items:
                    type: integer
                  description: Shuffled list of available bot profile-pic indices; each bot gets the next one (no duplicates).
                callbackUrl:
                  type: string
                  nullable: true
                  description: URL to POST to when the game finishes (empty body). Best-effort — failures are logged and ignored.
      responses:
        '200':
          description: Game was started
        '404':
          description: Game could not be found
        '409':
          description: Game was already started
    delete:
      summary: Stop a game
      operationId: stopGame
      tags:
        - games
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '204':
          description: Game was stopped
        '404':
          description: Game could not be found
  /games/{sessionId}/restart:
    post:
      summary: Restart a finished game with the same players but new game options
      description: |
        Resets the game state while keeping the same session ID, player IDs, and player names.
        Only valid when the game is in the FINISHED state.
      operationId: restartGame
      tags:
        - games
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                gameOptions:
                  type: object
                  additionalProperties: {}
      responses:
        '200':
          description: Game restarted successfully
        '404':
          description: Game not found
        '409':
          description: Game is not finished and cannot be restarted
  /games/{sessionId}/players:
    get:
      summary: Get all players in a game
      operationId: getAllPlayersInGame
      tags:
        - players
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: An array of all players in the session
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'
    post:
      summary: Add a player to a game
      operationId: addPlayerToGame
      tags:
        - players
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Player'
      responses:
        '201':
          description: Player was added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  playerId:
                    type: string
        '400':
          description: Player could not be added
        '404':
          description: Game not found
        '409':
          description: Game already started; no more players can join
  /games/{sessionId}/players/{playerId}:
    delete:
      summary: Remove a player from a game (player leaves voluntarily)
      operationId: leaveGame
      tags:
        - players
      parameters:
        - $ref: '#/components/parameters/sessionId'
        - $ref: '#/components/parameters/playerId'
      responses:
        '204':
          description: Player left the game
        '404':
          description: Game or player not found
...
  /gamestates/{sessionId}:
    get:
      summary: Get the game state for a specific game
      operationId: getGameState
      tags:
        - gamestates
      parameters:
        - $ref: '#/components/parameters/sessionId'
        - name: stateVersion
          in: query
          description: The version of the game state currently held by the client
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Game state (new version)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameState'
        '304':
          description: Game state not modified (client is up-to-date)
        '404':
          description: Game not found
  /games/{sessionId}/scores:
    get:
      summary: Get scores for all players
      description: Returns per-player score breakdown. Only available after game over.
      operationId: getScores
      tags:
        - games
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Map of playerId to ScoreCard
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/ScoreCard'
        '404':
          description: Game not found
        '409':
          description: Game not yet finished
...
components:
  parameters:
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
    playerId:
      name: playerId
      in: path
      required: true
      schema:
        type: string
  schemas:
    Color:
      type: string
      enum:
        - RED
        - YELLOW
        - GREEN
        - BLUE
    ClosureNotification:
      type: object
      required:
        - playerName
        - rowColor
      properties:
        playerName:
          type: string
          description: Name of the player who wants to close the row.
        rowColor:
          $ref: '#/components/schemas/Color'
    TurnPhase:
      type: string
      enum:
        - ROLL
        - ACTIVE_MOVE
        - PASSIVE_MOVE
        - EVALUATE
    MoveType:
      type: string
      enum:
        - ROLL
        - CROSS_WHITE_WHITE
        - CROSS_COLOR_DIE
        - CROSS_LOCK
        - TAKE_PUNISHMENT
        - PASS
        - DECLARE_LOCK_INTENT
        - RESET_TURN
        - GIVE_UP
        - UNDO_LAST_CROSS
        - CROSS_LUCKY_CROSS
    AvailableMove:
      type: object
      required:
        - cellId
        - moveType
      properties:
        cellId:
          type: string
        moveType:
          $ref: '#/components/schemas/MoveType'
    GameStatus:
      type: string
      enum:
        - waiting
        - inProgress
        - finished
    GameInfo:
      type: object
      required:
        - id
        - roomName
        - nrPlayers
        - maxNrPlayers
        - status
      properties:
        id:
          type: string
        roomName:
          type: string
        nrPlayers:
          type: integer
        maxNrPlayers:
          type: integer
        status:
          $ref: '#/components/schemas/GameStatus'
    Player:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        profilePic:
          type: string
          description: Zero-based global sprite index selected in the GameRoom lobby.
    NewGameRequest:
      type: object
      required:
        - roomName
        - maxPlayers
      properties:
        roomName:
          type: string
          minLength: 3
          maxLength: 50
        maxPlayers:
          type: integer
          minimum: 1
          maximum: 5
        gameOptions:
          type: object
          additionalProperties: {}
          description: Optional key-value map of game settings (e.g. {"cardMode":"SAME_CARDS"})
...
    RollResult:
      type: object
      description: The result of a dice roll.
      required:
        - white1
        - white2
        - coloredDice
      properties:
        white1:
          type: integer
          minimum: 1
          maximum: 6
        white2:
          type: integer
          minimum: 1
          maximum: 6
        coloredDice:
          type: object
          description: Map of color to die face value.
          additionalProperties:
            type: integer
            minimum: 1
            maximum: 6
    RowState:
      type: object
      required:
        - crossedCells
        - lockCrossed
      properties:
        crossedCells:
          type: array
          items:
            type: string
          description: IDs of crossed cells in this row.
        lockCrossed:
          type: boolean
    SheetProgress:
      type: object
      description: A player's crossing progress across all rows.
      required:
        - rowStates
        - punishments
      properties:
        rowStates:
          type: object
          description: Map of rowId to RowState.
          additionalProperties:
            $ref: '#/components/schemas/RowState'
        punishments:
          type: integer
          description: Number of punishment marks taken.
    TurnState:
      type: object
      required:
        - activePlayerId
        - phase
      properties:
        activePlayerId:
          type: string
        phase:
          $ref: '#/components/schemas/TurnPhase'
        currentRoll:
          $ref: '#/components/schemas/RollResult'
        passivePlayerQueue:
          type: array
          items:
            type: string
          description: Players still to complete their passive move this turn.
        whiteWhiteUsed:
          type: boolean
          description: Whether the active player has already used the white+white option this turn.
        colorDieUsed:
          type: boolean
          description: Whether the active player has already used the color die option this turn.
        luckyNumberUsed:
          type: boolean
          description: Whether the active player has used their Lucky Number move this turn.
        luckyCrossUsed:
          type: array
          items:
            type: string
          description: Player IDs that have already used their Lucky Cross bonus this turn.
        pendingCrosses:
          type: object
          description: Map of playerId to cell IDs crossed this turn but not yet confirmed. Empty once the turn ends.
          additionalProperties:
            type: array
            items:
              type: string
        effectiveWhiteWhite:
          type: object
          description: Map of playerId to effective white+white value after crossing an x-change cell this turn.
          additionalProperties:
            type: integer
    GameState:
      type: object
      required:
        - players
        - sheetProgress
        - sheetLayouts
        - gameOver
        - version
      properties:
        players:
          type: array
          items:
            $ref: '#/components/schemas/Player'
          description: Ordered list of players; defines turn order.
        sheetProgress:
          type: object
          description: Map of playerId to SheetProgress.
          additionalProperties:
            $ref: '#/components/schemas/SheetProgress'
        sheetLayouts:
          type: object
          description: Map of playerId to SheetLayout (static board structure).
          additionalProperties:
            $ref: '#/components/schemas/SheetLayout'
        closedRows:
          type: object
          description: Map of rowId to the playerId who closed it.
          additionalProperties:
            type: string
        activeDiceColors:
          type: array
          items:
            $ref: '#/components/schemas/Color'
          description: Colors of dice still in play (shrinks as rows are locked).
        turnState:
          $ref: '#/components/schemas/TurnState'
          description: Absent in offline mode.
        bonusNumbers:
          type: object
          description: Map of playerId to bonus numbers (Longo variant only; absent for standard games).
          additionalProperties:
            type: array
            items:
              type: integer
        closureNotifications:
          type: array
          items:
            $ref: '#/components/schemas/ClosureNotification'
          description: Pending row closure notifications for the UI.
        pendingClosures:
          type: object
          description: Map of rowId to the playerId who declared closing intent; rows in this map close at EVALUATE.
          additionalProperties:
            type: string
        availableMoves:
          type: object
          description: Map of playerId to available cross-cell moves for this state; absent in offline mode.
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/AvailableMove'
        seeOtherCards:
          type: boolean
          description: When false, clients hide the miniature sheets of other players. Defaults to true.
        doubleA:
          type: boolean
          description: Double A variant — every coloured cell has a twin cell below it. Clients render two-tier cells.
        doubleB:
          type: boolean
          description: Double B variant — some cells can be crossed twice. Clients render diagonal double cells.
        gameOver:
          type: boolean
        version:
          type: integer
          format: int64
          description: Increments on every state change; used by clients to detect updates.
...
    SheetCell:
      type: object
      required:
        - id
        - position
        - displayValue
        - color
        - closingEligible
        - tags
      properties:
        id:
          type: string
        position:
          type: integer
          description: Ordinal index in the row; engine uses this for progression checks.
        displayValue:
          type: string
          description: Value shown to the player (e.g. "7", "12").
        color:
          $ref: '#/components/schemas/Color'
        closingEligible:
          type: boolean
          description: Whether this cell can satisfy the lock's required-cells condition.
        tags:
          type: array
          items:
            $ref: '#/components/schemas/CellTag'
    LockConfig:
      type: object
      required:
        - id
        - color
        - minCrosses
        - closingCells
      properties:
        id:
          type: string
        color:
          $ref: '#/components/schemas/Color'
        minCrosses:
          type: integer
          description: Minimum number of crossed cells required before the lock can be triggered.
        closingCells:
          type: array
          items:
            type: string
          description: Cell IDs that can each independently close the row (any ONE is sufficient).
    SheetRow:
      type: object
      required:
        - id
        - cells
        - lock
      properties:
        id:
          type: string
        cells:
          type: array
          items:
            $ref: '#/components/schemas/SheetCell'
        lock:
          $ref: '#/components/schemas/LockConfig'
        bonusRow:
          type: boolean
          default: false
          description: >
            True when this row is a scoring bonus row (e.g. Big Points) that should
            not appear as a separate colour column in the score animation.
            Its points are already captured in the dedicated bonusPoints field of ScoreCard.
        luckyRow:
          type: boolean
          default: false
          description: True when this row is the Lucky Number bonus row.
        bonusBar:
          type: boolean
          default: false
          description: True when this row is the Bonus A bonus bar (12 coloured cells below the four colour rows).
        luckyTarget:
          type: integer
          description: >
            For lucky rows only: the target dice sum (white1 + white2 + any coloured die)
            that unlocks a cross. 15 for Standard, 20 for Longo. Absent on non-lucky rows.
        upperNeighbourRowId:
          type: string
          description: >
            For bonus rows only: the ID of the upper neighbouring colour row.
            A bonus cell may only be played when its display value is crossed in at least
            one of the two neighbouring rows.
        lowerNeighbourRowId:
          type: string
          description: >
            For bonus rows only: the ID of the lower neighbouring colour row.
    SheetLayout:
      type: object
      required:
        - rows
      properties:
        rows:
          type: array
          items:
            $ref: '#/components/schemas/SheetRow'
    ScoreCard:
      type: object
      required:
        - crossesPerColor
        - pointsPerColor
        - extraCrosses
        - extraPoints
        - bonusPoints
        - punishmentPoints
        - total
      properties:
        crossesPerColor:
          type: object
          description: Number of crossed cells per color.
          additionalProperties:
            type: integer
        pointsPerColor:
          type: object
          description: Triangular score per color.
          additionalProperties:
            type: integer
        extraCrosses:
          type: integer
          description: Crossed cells tagged ExtraBucket.
        extraPoints:
          type: integer
          description: Triangular score from extra crosses.
        bonusPoints:
          type: integer
          description: Flat bonus points from BonusPoints tags.
        punishmentPoints:
          type: integer
          description: Negative points from punishments (-5 each).
        total:
          type: integer
          description: Sum of all points.
...